Move symbol status tracking to the refer pass.
This substantially simplifies the work done in the expr pass,
and isolates the symbol status tracking to a much simpler pass itself.
