+++ src/libponyc/ast/token.h
@@ -135,6 +135,7 @@ typedef enum token_id
   TK_CAP_ANY,
 
   TK_THIS,
+  TK_THISINCOMPLETE,
   TK_RETURN,
   TK_BREAK,
   TK_CONTINUE,
