+++ src/libponyc/expr/call.h
@@ -7,8 +7,6 @@
 
 PONY_EXTERN_C_BEGIN
 
-bool is_this_incomplete(typecheck_t* t, ast_t* ast);
-
 bool expr_call(pass_opt_t* opt, ast_t** astp);
 
 PONY_EXTERN_C_END
