+++ src/libponyc/expr/control.h
@@ -14,7 +14,6 @@ bool expr_repeat(pass_opt_t* opt, ast_t* ast);
 bool expr_try(pass_opt_t* opt, ast_t* ast);
 bool expr_recover(pass_opt_t* opt, ast_t* ast);
 bool expr_break(pass_opt_t* opt, ast_t* ast);
-bool expr_continue(pass_opt_t* opt, ast_t* ast);
 bool expr_return(pass_opt_t* opt, ast_t* ast);
 bool expr_error(pass_opt_t* opt, ast_t* ast);
 bool expr_compile_error(pass_opt_t* opt, ast_t* ast);
