+++ src/libponyc/pass/expr.c
@@ -228,7 +228,6 @@ ast_result_t pass_expr(ast_t** astp, pass_opt_t* options)
     case TK_VAR:
     case TK_LET:        r = expr_local(options, ast); break;
     case TK_BREAK:      r = expr_break(options, ast); break;
-    case TK_CONTINUE:   r = expr_continue(options, ast); break;
     case TK_RETURN:     r = expr_return(options, ast); break;
     case TK_IS:
     case TK_ISNT:       r = expr_identity(options, ast); break;
@@ -261,12 +260,10 @@ ast_result_t pass_expr(ast_t** astp, pass_opt_t* options)
     case TK_LETREF:     r = expr_localref(options, ast); break;
     case TK_PARAMREF:   r = expr_paramref(options, ast); break;
 
+    case TK_THISINCOMPLETE:
     case TK_THIS:       r = expr_this(options, ast); break;
     case TK_TRUE:
     case TK_FALSE:      r = expr_literal(options, ast, "Bool"); break;
-    case TK_ERROR:      r = expr_error(options, ast); break;
-    case TK_COMPILE_ERROR:
-                        r = expr_compile_error(options, ast); break;
     case TK_COMPILE_INTRINSIC:
                         r = expr_compile_intrinsic(options, ast); break;
     case TK_LOCATION:   r = expr_location(options, ast); break;
